Make clnt independent of sbuf2 (#2569)
