+++ db/verify.h
@@ -17,8 +17,6 @@
 #ifndef INCLUDED_VERIFY_H
 #define INCLUDED_VERIFY_H
 
-struct dbtable;
-
 typedef enum {
     VERIFY_SERIAL,
     VERIFY_PARALLEL,
@@ -26,14 +24,15 @@ typedef enum {
     VERIFY_INDICES,
     VERIFY_BLOBS
 } verify_mode_t;
-void purge_by_genid(struct dbtable *db, unsigned long long *genid);
-void dump_record_by_rrn_genid(struct dbtable *db, int rrn, unsigned long long genid);
-int verify_table(const char *table, SBUF2 *sb, int progress_report_seconds,
-                 int attempt_fix, int (*lua_callback)(void *, const char *),
-                 void *lua_params);
-int verify_table_mode(const char *table, SBUF2 *sb, int progress_report_seconds,
-                      int attempt_fix,
-                      int (*lua_callback)(void *, const char *),
-                      void *lua_params, verify_mode_t mode);
+
+struct dbtable;
+void purge_by_genid(struct dbtable *, unsigned long long *genid);
+void dump_record_by_rrn_genid(struct dbtable *, int rrn, unsigned long long genid);
+
+typedef int(verify_peer_check_func)(void *arg);
+typedef int(verify_response_func)(char *response, void *arg);
+int verify_table(const char *table, int progress_report_seconds,
+                 int attempt_fix, verify_mode_t, verify_peer_check_func *,
+                 verify_response_func *, void *arg);
 
 #endif
