+++ db/watchdog.c
@@ -146,16 +146,12 @@ int gbl_epoch_time; /* db has been up gbl_epoch_time - gbl_starttime seconds */
 
 static void watchdogsql(void)
 {
-    struct sqlclntstate clnt = {0};
-    int fd = open("/dev/null", O_RDWR);
-    SBUF2 *sb = sbuf2open(fd, 0);
+    struct sqlclntstate clnt;
     start_internal_sql_clnt(&clnt);
     clnt.dbtran.mode = TRANLEVEL_SOSQL;
-    clnt.sb = sb;
     clnt.admin = 1;
     run_internal_sql_clnt(&clnt, "select 1");
     end_internal_sql_clnt(&clnt);
-    sbuf2close(sb);
 }
 
 static void *watchdog_thread(void *arg)
