+++ lua/sp.h
@@ -37,6 +37,5 @@ int is_pingpong(struct sqlclntstate *);
 void lua_final(struct sqlite3_context *);
 void lua_step(struct sqlite3_context *, int argc, struct sqlite3_value **argv);
 void lua_func(struct sqlite3_context *, int argc, struct sqlite3_value **argv);
-int db_verify_table_callback(void *v, const char *buf);
 
 #endif
