+++ util/CMakeLists.txt
@@ -26,6 +26,7 @@ set(src
   nodemap.c
   object_pool.c
   parse_lsn.c
+  pb_alloc.c
   perf.c
   plhash.c
   pool.c
@@ -63,18 +64,21 @@ if(COMDB2_LEGACY_DEFAULTS)
 endif()
 add_library(util ${src})
 include_directories(
-  ${CMAKE_CURRENT_SOURCE_DIR}
   ${CMAKE_CURRENT_BINARY_DIR}
+  ${CMAKE_CURRENT_SOURCE_DIR}
+  ${PROJECT_BINARY_DIR}/mem
+  ${PROJECT_BINARY_DIR}/protobuf
   ${PROJECT_SOURCE_DIR}/bbinc
+  ${PROJECT_SOURCE_DIR}/berkdb
   ${PROJECT_SOURCE_DIR}/cdb2api
+  ${PROJECT_SOURCE_DIR}/db
   ${PROJECT_SOURCE_DIR}/dlmalloc
   ${PROJECT_SOURCE_DIR}/mem
-  ${PROJECT_BINARY_DIR}/mem
+  ${PROJECT_SOURCE_DIR}/protobuf
   ${PROJECT_SOURCE_DIR}/sockpool
-  ${PROJECT_SOURCE_DIR}/berkdb
+  ${PROTOBUF-C_INCLUDE_DIR}
   ${OPENSSL_INCLUDE_DIR}
   ${UNWIND_INCLUDE_DIR}
-  ${PROJECT_SOURCE_DIR}/db
 )
 add_dependencies(util mem)
 set_source_files_properties(walkback.c PROPERTIES COMPILE_FLAGS -DUSE_UNWIND)
